test: update helper functions
